Move mocks into their own module
