+++ modules/src/ics02_client/client.rs
@@ -2,12 +2,16 @@
 
 use crate::ics02_client::state::{ClientState, ConsensusState};
 
-pub trait ClientDef: Sized {
+pub trait ClientDef {
     type Error: std::error::Error;
 
     type Header;
-    type ClientState: ClientState + From<Self::ConsensusState>;
+    type ClientState: ClientState;
     type ConsensusState: ConsensusState;
+
+    fn new_client_state(_consensus_state: &Self::ConsensusState) -> Self::ClientState {
+        todo!()
+    }
 
     /// TODO
     fn check_validity_and_update_state(
