+++ modules/src/ics02_client/mod.rs
@@ -6,6 +6,7 @@ pub mod error;
 pub mod events;
 pub mod handler;
 pub mod header;
+pub mod mocks;
 pub mod msgs;
 pub mod raw;
 pub mod state;
